rename() : repackage - view : Printer, Scanner
